+++ src/main/java/com/proyectosgrt/demo/repository/Tiposdoc_repository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.proyectosgrt.demo.models.Tiposdoc;
 
-public interface Tiposdoc_repository extends JpaRepository <Tiposdoc, Long> {
+public interface Tiposdoc_repository extends JpaRepository <Tiposdoc, Byte> {
 
 }
